Refactor architecture into cogs, core, and services
- Moved core logic to `core/` (models, config, storage, ui, parallel_group_creator)
- Extracted utility functions to `core/utils.py`
- Created `services/group_service.py` to handle group creation state and logic
- Split commands into `cogs/general.py`, `cogs/roles.py`, `cogs/groups.py`, `cogs/debug.py`
- Simplified `bot.py` to load extensions and initialize services
- Updated tests to reflect new structure
