+++ tests/test_group_creator.py
@@ -5,8 +5,8 @@ import unittest
 # Add the parent directory to the Python path so we can import our modules
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from models import WoWPlayer
-from parallel_group_creator import clear, create_mythic_plus_groups
+from core.models import WoWPlayer
+from core.parallel_group_creator import clear, create_mythic_plus_groups
 from tests.prebuilt_classes import (
     BalanceDruid,
     FeralDruid,
