+++ tests/test_issues.py
@@ -3,16 +3,16 @@ from unittest.mock import AsyncMock, patch
 
 import discord
 
-from issues import GitHubError, GitHubIssueModal, create_github_issue
+from core.issues import GitHubError, GitHubIssueModal, create_github_issue
 
 
 class TestIssues(unittest.IsolatedAsyncioTestCase):
     async def test_create_github_issue_success(self):
         # Mock config values
         with (
-            patch("config.GITHUB_TOKEN", "fake_token"),
-            patch("config.GITHUB_REPO_OWNER", "owner"),
-            patch("config.GITHUB_REPO_NAME", "repo"),
+            patch("core.config.GITHUB_TOKEN", "fake_token"),
+            patch("core.config.GITHUB_REPO_OWNER", "owner"),
+            patch("core.config.GITHUB_REPO_NAME", "repo"),
         ):
             with patch("aiohttp.ClientSession.post") as mock_post:
                 mock_response = AsyncMock()
@@ -27,9 +27,9 @@ class TestIssues(unittest.IsolatedAsyncioTestCase):
 
     async def test_create_github_issue_failure(self):
         with (
-            patch("config.GITHUB_TOKEN", "fake_token"),
-            patch("config.GITHUB_REPO_OWNER", "owner"),
-            patch("config.GITHUB_REPO_NAME", "repo"),
+            patch("core.config.GITHUB_TOKEN", "fake_token"),
+            patch("core.config.GITHUB_REPO_OWNER", "owner"),
+            patch("core.config.GITHUB_REPO_NAME", "repo"),
         ):
             with patch("aiohttp.ClientSession.post") as mock_post:
                 mock_response = AsyncMock()
@@ -57,7 +57,7 @@ class TestIssues(unittest.IsolatedAsyncioTestCase):
         modal.extra_info._value = "Steps"  # pyright: ignore[reportPrivateUsage]
 
         # Mock config and create_github_issue
-        with patch("issues.create_github_issue", new_callable=AsyncMock) as mock_create:
+        with patch("core.issues.create_github_issue", new_callable=AsyncMock) as mock_create:
             mock_create.return_value = {"html_url": "http://url"}
 
             await modal.on_submit(mock_interaction)
